refactor: decouple demos from graph_traversala
