+++ src/graph_traversals/dijkstra.c
@@ -417,217 +417,3 @@ void free_weightedGraph(weightedGraph* graph)
     free(graph->array);
     free(graph);
 }
-
-void dijkstra_demo(void)
-{
-    int edges;
-    int graph_capacity;
-    int starting_node;
-    int input_method;
-    weightedGraph* graph = NULL;
-
-    while (1)
-    {
-        int method_status = safe_input_int(&input_method,
-                                           "\nenter 1 to build the graph manually, 2 to load it "
-                                           "from a CSV file, enter '-1' to exit : ",
-                                           1, 2);
-
-        if (method_status == INPUT_EXIT_SIGNAL)
-        {
-            printf("\nExiting Dijkstra demo.....\n");
-            return;
-        }
-
-        if (method_status == 0)
-        {
-            continue;
-        }
-
-        break;
-    }
-
-    if (input_method == 2)
-    {
-        while (1)
-        {
-            char path[256];
-            printf("\nenter the path to the CSV file, enter '-1' to exit : ");
-            fflush(stdout);
-
-            if (!fgets(path, sizeof(path), stdin))
-            {
-                printf("\ninput ended unexpectedly\n");
-                clearerr(stdin);
-                return;
-            }
-
-            size_t len = strlen(path);
-            while (len > 0 && (path[len - 1] == '\n' || path[len - 1] == '\r'))
-                path[--len] = '\0';
-
-            if (strcmp(path, "-1") == 0)
-            {
-                printf("\nExiting Dijkstra demo.....\n");
-                return;
-            }
-
-            if (len == 0)
-            {
-                continue;
-            }
-
-            graph = load_weightedGraph_from_csv(path);
-
-            if (!graph)
-            {
-                // error already reported by the loader, let the user retry
-                continue;
-            }
-
-            break;
-        }
-
-        graph_capacity = graph->V;
-    }
-    else
-    {
-        while (1)
-        {
-            int graph_capacity_status =
-                safe_input_int(&graph_capacity,
-                               "\nenter the number of vertices in the graph, "
-                               "(between 1 and 10), enter '-1' to exit : ",
-                               1, 10);
-
-            if (graph_capacity_status == INPUT_EXIT_SIGNAL)
-            {
-                printf("\nExiting Dijkstra demo.....\n");
-                free_weightedGraph(graph);
-                return;
-            }
-
-            if (graph_capacity_status == 0)
-            {
-                continue;
-            }
-
-            graph = create_weightedGraph(graph_capacity);
-
-            if (!graph)
-            {
-                printf("\nmemory allocation failed\n");
-                free_weightedGraph(graph);
-                return;
-            }
-
-            break;
-        }
-
-        while (1)
-        {
-            int edges_capacity_status = safe_input_int(
-                &edges, "\nenter number of edges (between 1 and 100), enter '-1' to exit :", 0,
-                100);
-
-            if (edges_capacity_status == INPUT_EXIT_SIGNAL)
-            {
-                printf("\nExiting Dijkstra demo\n");
-                free_weightedGraph(graph);
-                return;
-            }
-
-            if (edges_capacity_status == 0)
-            {
-                continue;
-            }
-
-            break;
-        }
-
-        printf(
-            "\nEnter source, destination, weight pairs (Source, Destination must be b/w 0 and %d "
-            "(both inclusive)):\n",
-            graph_capacity - 1);
-
-        for (int i = 0; i < edges; i++)
-        {
-            int src_status;
-            int dest_status;
-            int wt_status;
-            int src;
-            int dest;
-            int wt;
-
-        retry:
-            src_status = safe_input_int(&src, "src: ", 0, graph_capacity - 1);
-
-            if (src_status == INPUT_EXIT_SIGNAL)
-            {
-                printf("\nExiting Dijkstra demo\n");
-                free_weightedGraph(graph);
-                return;
-            }
-            if (src_status == 0)
-            {
-                goto retry;
-            }
-
-            dest_status = safe_input_int(&dest, "dest: ", 0, graph_capacity - 1);
-
-            if (dest_status == INPUT_EXIT_SIGNAL)
-            {
-                printf("\nExiting Dijkstra demo\n");
-                free_weightedGraph(graph);
-                return;
-            }
-            if (dest_status == 0)
-            {
-                goto retry;
-            }
-
-            wt_status = safe_input_int(&wt, "weight: ", 0, INT_MAX);
-
-            if (wt_status == INPUT_EXIT_SIGNAL)
-            {
-                printf("\nExiting Dijkstra demo\n");
-                free_weightedGraph(graph);
-                return;
-            }
-            if (wt_status == 0)
-            {
-                goto retry;
-            }
-
-            add_edge_directed(graph, src, dest, wt);
-        }
-    }
-
-    while (1)
-    {
-        int start_status =
-            safe_input_int(&starting_node, "\nenter starting node: ", 0, graph_capacity - 1);
-
-        if (start_status == INPUT_EXIT_SIGNAL)
-        {
-            printf("\nExiting Dijkstra demo.....\n");
-            free_weightedGraph(graph);
-            return;
-        }
-
-        if (start_status == 0)
-            continue;
-
-        if (starting_node < 0 || starting_node >= graph->V)
-        {
-            printf("Invalid start node\n");
-            free_weightedGraph(graph);
-            return;
-        }
-        break;
-    }
-
-    printf("\n");
-    dijkstra(graph, starting_node);
-    free_weightedGraph(graph);
-}
